Expose value type to data entity
The value type is very helpful in cases where one only
has the abstract DataEntity at hand, but have to handle
the actual type accordingly (e.g. for serializing) and where
the observationType is not at hand.

The value type has been always just the discriminator value
for different observation types so it is just mapped to the
data entity as non-updatable and non-insertable.